+++ src/routes/auth.js
@@ -1,0 +1,8 @@
+'use strict';
+
+const _express = require('express');
+const _router = _express.Router();
+
+const ctrl = require('../controllers/auth');
+
+module.exports = _router;
