+++ src/routes/menu.js
@@ -3,6 +3,6 @@
 const _express = require('express');
 const _router = _express.Router();
 
-const ctrl = require('./service');
+const ctrl = require('../controllers/menu');
 
 module.exports = _router;
